Update yt-heist.sh - batchfiles
using the script in default mode now makes use of yt-dlp's -a (batchfile) option. this overcomes a severe issue when downloading YT videos from just a list of IDs. If an ID begins with a hyphen (-), yt-dlp treats it as a CLI flag! -a fixes this, and also the script saves the IDs of the last session for easy debugging